refactor: move icon component to common UiKit page
