+++ @ui/ui-kit/Icon/IconsComponent.module.scss
@@ -1,11 +1,11 @@
-@use '../../../styles/library';
+@use '../../design/settings/colors.scss';
 
 .IconWrapper {
 
   display: flex;
   transition: .2s;
   width: fit-content;
-  color: library.$light-100;
+  color: colors.$light-100;
   box-sizing: content-box;
   text-decoration: none;
 
@@ -14,7 +14,7 @@
   }
 
   &:hover {
-    color: library.$primary-100;
+    color: colors.$primary-100;
     cursor: pointer;
 
     .Icon {
@@ -25,7 +25,7 @@
   &:active,
   &Active,
   &Active:hover {
-    color: library.$primary-500;
+    color: colors.$primary-500;
 
     .Icon {
       filter: invert(36%) sepia(80%) saturate(1143%) hue-rotate(197deg) brightness(102%) contrast(93%);
@@ -33,14 +33,13 @@
   }
 
   &:focus {
-    outline: 2px solid library.$primary-100;
+    outline: 2px solid colors.$primary-100;
     border-radius: 4px;
   }
 
   &Disabled,
   &Disabled:hover {
-    color: library.$dark-100;
-    outline: 0px solid black;
+    color: colors.$dark-100;
     cursor: not-allowed;
 
     .Icon {
@@ -49,7 +48,6 @@
   }
 
   .Text {
-    margin-top: 0px;
     margin-left: 12px;
     font-size: 14px;
     font-weight: 500;
@@ -66,7 +64,6 @@
   }
 
   .Text {
-    margin-top: 0px;
     margin-left: 12px;
     font-size: 14px;
     font-weight: 500;
