+++ @ui/ui-kit/Icon/iconsType.ts
