+++ pages/dev/ui-kit/index.tsx
@@ -3,6 +3,7 @@ import { getLayout } from '@/components/Layout/Layout'
 import { Button } from '@/@ui/ui-kit/Button/Button'
 import style from './index.module.scss'
 import { BUTTON_COLORS } from '@/@ui/ui-kit/Button/constants'
+import IconsDev from '@/components/dev/UiKit/Icons'
 
 export const getStaticProps = async () => {
   return {
@@ -17,32 +18,33 @@ const Login = () => {
         <h1 className={style.header}> UIKit Page</h1>
 
         <div className={style.kitBlock}>
-            <h2 className={style.componentHeader}>Buttons</h2>
-            <div className={style.components}>
-
-                <div className={style.buttonBlock}>
-                <Button text='Button'></Button>
-                <Button text='Button' disabled></Button>
-                </div>
-
-                <div className={style.buttonBlock}>
-                <Button color={BUTTON_COLORS.BASIC} text='Button'></Button>
-                <Button color={BUTTON_COLORS.BASIC} text='Button' disabled></Button>
-                </div>
-
-                <div className={style.buttonBlock}>
-                <Button color={BUTTON_COLORS.OUTLINED} text='Button'></Button>
-                <Button color={BUTTON_COLORS.OUTLINED} text='Button' disabled></Button>
-                </div>
-
-                <div className={style.buttonBlock}>
-                <Button color={BUTTON_COLORS.GHOST} text='Button'></Button>
-                <Button color={BUTTON_COLORS.GHOST} text='Button' disabled></Button>
-                </div>
-
+          <h2 className={style.componentHeader}>Buttons</h2>
+          <div className={style.components}>
+            <div className={style.buttonBlock}>
+              <Button text="Button"></Button>
+              <Button text="Button" disabled></Button>
             </div>
-        </div>
 
+            <div className={style.buttonBlock}>
+              <Button color={BUTTON_COLORS.BASIC} text="Button"></Button>
+              <Button color={BUTTON_COLORS.BASIC} text="Button" disabled></Button>
+            </div>
+
+            <div className={style.buttonBlock}>
+              <Button color={BUTTON_COLORS.OUTLINED} text="Button"></Button>
+              <Button color={BUTTON_COLORS.OUTLINED} text="Button" disabled></Button>
+            </div>
+
+            <div className={style.buttonBlock}>
+              <Button color={BUTTON_COLORS.GHOST} text="Button"></Button>
+              <Button color={BUTTON_COLORS.GHOST} text="Button" disabled></Button>
+            </div>
+          </div>
+        </div>
+        <div className={style.kitBlock}>
+          <h2 className={style.componentHeader}>Buttons</h2>
+          <IconsDev />
+        </div>
       </div>
     </PageWrapper>
   )
